feat: S12P21A601-70 [BE] 약속방 생성 및 참여 기능 구현 코드 리뷰 반영

+++ back-end/moba/src/main/java/com/a601/moba/appointment/Repository/AppointmentParticipantRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentParticipantRepository extends JpaRepository<AppointmentParticipant, Long> {
+public interface AppointmentParticipantRepository extends JpaRepository<AppointmentParticipant, Integer> {
     List<AppointmentParticipant> findAllByAppointment(Appointment appointment);
 
     Optional<AppointmentParticipant> findByAppointmentAndMemberId(Appointment appointment, Integer memberId);
