+++ back-end/moba/src/main/java/com/a601/moba/appointment/Repository/AppointmentRepository.java
@@ -3,6 +3,6 @@ package com.a601.moba.appointment.Repository;
 import com.a601.moba.appointment.Entity.Appointment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
 
 }
